requirements.txt
flask
psycopg2-binary
python-dotenv
sentence-transformers
rank-bm25
groq
numpy
pgvector